SSL Websocket Fix (#2165)
* SSL Websocket Fix

* static_cast<size_t>

* Attempt GCC fix

* dispatch helpers

* Use lowest_layer everywhere

* wss server support

* Update http_client_test.cpp

* fixes

* More clear CI failures for SSL testing

* Update websocket_client.hpp

+++ include/glaze/rpc/repe/buffer.hpp
@@ -237,7 +237,7 @@ namespace glz::repe
          return {};
       }
 
-      return {data + sizeof(header), hdr.query_length};
+      return {data + sizeof(header), static_cast<size_t>(hdr.query_length)};
    }
 
    /// Extract just the query string from string_view
